requirements.txt
aiogram==3.13.1
supabase-py==2.4.0
python-dotenv==1.0.1
